requirements.txt
django
djangorestframework
drf-spectacular
django-auth-ldap
djangorestframework-simplejwt
django-cors-headers